# perf: add rax_tree_up() to iterate the routes tree