--- a/benchmark/match-prefix.lua
+++ b/benchmark/match-prefix.lua
@@ -1,6 +1,6 @@
 local radix = require("resty.radixtree")
 local route_count = 1000 * 100
-local match_times = 1000 * 1
+local match_times = 1000 * 1000
 
 local routes = {}
 for i = 1, route_count do