--- a/src/easy_rax.h
+++ b/src/easy_rax.h
@@ -57,6 +57,7 @@ void *radix_tree_find(void *t, const unsigned char *buf, size_t len);
 void *radix_tree_search(void *t, void *it, const unsigned char *buf, size_t len);
 int radix_tree_prev(void *it, const unsigned char *buf, size_t len);
 int radix_tree_next(void *it, const unsigned char *buf, size_t len);
+int radix_tree_up(void *it, const unsigned char *buf, size_t len);
 int radix_tree_stop(void *it);
 
 void *radix_tree_new_it(void *t);