@haoyang1994 2024-01-12
@membphis  Hi, this is my first time using CLA. Does this mean I have signed? But it looks that the CLA checking is still pending.
![image](https://github.com/api7/lua-resty-radixtree/assets/17779987/b0c0c1ca-f776-4ce5-9bec-d8d679348fb7)